Lighting range/attenuation fixup wip
In addition to the bitmask we only process lights we have actually
submitted that frame.  This is a half-measure before storing lights in
separate sparse arrays so that we can just iterate the bits set in the
mask.

Conical lights now have their own section carved out in DrawObb.

Fixup the gl error scope function so that it propagates the file/line
information.

Increase the range on the spotlight.

Add a note to TODO so that I remember what I'm doing here.