Initial imgui support, move around some files
Added working support for imgui, but with test debug menus.  Also
shuffled around files so that lodepng is now an External source.  Had to
add a few function entry points for GL functions that imgui requires.

Also removed the x86 and Win32 configurations from the solution and
project file respectively.